+++ test/unit/fuzz-tools.test.js
@@ -384,6 +384,7 @@ import * as reganMorreyV404 from '../../lib/regan-morrey-v404.js';
 import * as savaryMillerV405 from '../../lib/savary-miller-v405.js';
 import * as leFortV406 from '../../lib/le-fort-v406.js';
 import * as steinbergAvnV407 from '../../lib/steinberg-avn-v407.js';
+import * as meyersMckeeverV408 from '../../lib/meyers-mckeever-v408.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -713,6 +714,7 @@ const MODULES = {
   'savary-miller-v405.js': savaryMillerV405,
   'le-fort-v406.js': leFortV406,
   'steinberg-avn-v407.js': steinbergAvnV407,
+  'meyers-mckeever-v408.js': meyersMckeeverV408,
   'tb-testing.js': tbTesting,
 };